+++ deltafi-auth/models/cert_user.rb
@@ -20,6 +20,7 @@
 
 class CertUser
   include Deltafi::EntityResolver
+  include Deltafi::MetricsRoleResolver
 
   attr_reader :dn, :identifiers